+++ packages/typir/src/graph/type-node.ts
@@ -8,7 +8,6 @@ import { TypeReference } from '../initialization/type-reference.js';
 import { WaitingForIdentifiableAndCompletedTypeReferences, WaitingForInvalidTypeReferences } from '../initialization/type-waiting.js';
 import { Kind } from '../kinds/kind.js';
 import { TypirSpecifics } from '../typir.js';
-import { TypirProblem } from '../utils/utils-definitions.js';
 import { assertTrue, assertUnreachable, removeFromArray } from '../utils/utils.js';
 import { TypeEdge } from './type-edge.js';
 
@@ -324,15 +323,6 @@ export abstract class Type {
         }
     }
 
-
-
-    /**
-     * Analyzes, whether two types are equal.
-     * @param otherType to be compared with the current type
-     * @returns an empty array, if both types are equal, otherwise some problems which might point to found differences/conflicts between the two types.
-     * These problems are presented to users in order to support them with useful information about the result of this analysis.
-     */
-    abstract analyzeTypeEqualityProblems(otherType: Type): TypirProblem[];
 
 
     addIncomingEdge(edge: TypeEdge): void {