+++ packages/typir/src/kinds/class/top-class-type.ts
@@ -44,7 +44,7 @@ export class TopClassType extends Type implements TypeGraphListener {
         return this.getIdentifier();
     }
 
-    override analyzeTypeEqualityProblems(otherType: Type): TypirProblem[] {
+    protected analyzeTypeEqualityProblems(otherType: Type): TypirProblem[] {
         if (isTopClassType(otherType)) {
             return [];
         } else {